+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.4.11 - 12 Oct 2025
 Bug fix:
 - [#1469](https://github.com/elysiajs/elysia/issues/1469) incorrect ping, pong type signature
+- [#1467](https://github.com/elysiajs/elysia/issues/1467) better error union handling in `onError`
 - make `file-type` non optional dependency to fix default build problem
 
 # 1.4.10 - 9 Oct 2025